Fast Disjoint
Before, we were calculating whether or not two sets are disjoint by
performing x.intersect(y).nonEmpty. In Scala, this is equivalent to
performing x.filter(y).nonEmpty. In other words, the entire sets x and y
must be traversed in order to check whether or not they overlap. In
cases where the sets do overlap, this should have no impact on
performance. It should, however, improve performance when sets are not
disjoint.